cpp05

+++ Module04/ex02/Brain.hpp
@@ -6,7 +6,7 @@
 /*   By: mregrag <mregrag@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/28 21:10:08 by mregrag           #+#    #+#             */
-/*   Updated: 2024/11/28 21:54:32 by mregrag          ###   ########.fr       */
+/*   Updated: 2024/12/11 23:43:21 by mregrag          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,16 +18,16 @@
 
 class Brain
 {
-	private:
-		std::string ideas[100];
-	public:
-		Brain();
-		Brain(const Brain& rhs);
-		Brain& operator=(const Brain& rhs);
-		~Brain();
+    private:
+	std::string ideas[100];
+    public:
+	Brain();
+	Brain(const Brain& rhs);
+	Brain& operator=(const Brain& rhs);
+	~Brain();
 
-		void setIdea(size_t index, const std::string& idea);
-		std::string getIdea(size_t index) const;
+	void setIdea(int index, const std::string& idea);
+	std::string getIdea(int index) const;
 };
 
 #endif
